Http Server: register the models on call to their init function
This adds a description parameter to the module auto-generated init
function and use that description to register the API when calling to
it's init.

This also change all the init calls to include their description.

To check either call:
http://localhost:8000/api-doc/ to get the list of APIs and for example
to get the env api call:
http://localhost:8000/api-doc/env/env

or use inside the swagger-ui gui the url:
http://localhost:8000/api-doc

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>