Http server Tests: Compile a swagger client
This adds an optional compilation of the swagger client using the
swagger-codegen.

It downloads and compile the library from github, update the scala
version inside the build.sbt to 2.11.2.

The swagger-codegen parser runs against the server, so it is added
as a step to the tests.

To run the client generation as part of the testing use:
make validate-api check

By default the codegen is not running, it can be changed in the future.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>