Console Drivers: Add a memory console
This adds a console that will be used by the API, it holds both output and
input in memory.  The implementation is based on a circular buffer, so if no
text will be emptied from it, it will be over written.

When reading output from the console it will be removed, supporting one
connection at a time.

The input would go to any application that would look for input, i.e.
interactive app like the on-image CLI.

For this phase, the console numbers are hard coded, but it can be modified if
needed in the future.